Extract operation state and pass exceptionresource
